refactor: split src/tools/macos.ts into domain-specific files
Closes #73
